dom/base/test/test_bug626262.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=626262
-->
<head>
  <title>Test for Bug 626262</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=626262">Mozilla Bug 626262</a>
<p id="display"><iframe id="f" srcdoc="1"></iframe></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 626262 **/
SimpleTest.waitForExplicitFinish();

addLoadEvent(function() {
  var iframe = document.getElementById("f");
  var frameDoc = iframe.contentDocument;
  var parent = frameDoc.createElementNS("http://www.w3.org/1999/xhtml", "div");

  function a()
  {
    window.removeEventListener("DOMNodeRemoved", arguments.callee);
    document.adoptNode(parent);
  }

  var text = document.createTextNode(" ");
  document.documentElement.appendChild(text);

  var thrown = false;
  try {
    window.addEventListener("DOMNodeRemoved", a);
    parent.appendChild(text);
  }
  catch (e) {
    thrown = true;
  }

  ok(!thrown, "changing ownerDocument during adoptNode should not throw");

  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
